correcciones de prueba

+++ src/main/java/com/utc/proyectoF/repository/CargoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import com.utc.proyectoF.entity.Cargo;
 
 @Repository
-public interface CargoRepository extends JpaRepository<Cargo, Long> {
+public interface CargoRepository extends JpaRepository<Cargo, Integer> {
 
 }
